feat: add logic for random city button to login-page
